+++ src/filter_h1_T.ts
@@ -1,1 +1,0 @@
-export interface filter_h1_T extends Record<string, boolean> {}