+++ src/filter_r_T.ts
@@ -1,0 +1,1 @@
+export interface filter_r_T extends Record<string, boolean> {}