+++ src/index.ts
@@ -1,3 +1,4 @@
-export * from './filter_h1_'
-export * from './filter_h_2'
-export * from './filter_h1_T'
+export * from './filter_r_'
+export * from './filter_r_2'
+export * from './filter_r_T'
+export * from './r_a_present'